pyproject.toml
[project]
name = "safe-benchmark"
version = "0.1.0"
description = "SAFE framework benchmark experiment over τ³-bench tasks"
requires-python = ">=3.12,<3.14"
readme = "README.md"

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
